+++ example/update_statements.dart
@@ -1,0 +1,12 @@
+// ignore_for_file: cascade_invocations
+
+import 'package:database_query_builder/query_builder.dart';
+
+class UpdateStatements {
+  static Future<void> updateNotWhere() async {
+    final table = DB.table('people').update({'age': 15});
+    table.toSQL();
+    await table.save();
+    await table.where('age', 50).save();
+  }
+}