+++ lib/src/exec.dart
@@ -1,11 +1,9 @@
+import 'package:database_query_builder/src/data.dart';
 import 'package:database_query_builder/src/sql_query.dart';
-import 'package:database_query_builder/src/where.dart';
 
-class Exec extends Where<Exec> {
-  String toSQL() {
-    return SQLquery.instance.getSQL();
-  }
+class Execute with Exec, DataSQL {}
 
+class Exec {
   Future<int> save() async {
     return SQLquery.instance.executeSQL();
   }