+++ lib/src/sql_query.dart
@@ -19,17 +19,23 @@ class SQLquery {
   String table = '';
   List<String> query = [];
   Map<String, dynamic> params = {};
+  List<String> selects = [];
   String get paramsCode => 'p${params.keys.length}';
   DBconnect? dbconnect;
 
   String getSQL() {
-    final sql = query.join(' ');
-    final queryResult = PostgreSQLFormat.substitute(sql, params);
-    print('==============================================================>');
-    print('1) QUERY: $sql');
-    print('2) PARAMETER: $params');
-    print('3) SQL EXECUTE QUERY: $queryResult');
-    return queryResult;
+    try {
+      validateEnpty();
+
+      final sql = query.join(' ');
+      final queryResult = PostgreSQLFormat.substitute(sql, params);
+      selects.clear();
+      print('SQL: $queryResult');
+      return queryResult;
+    } catch (e) {
+      final sql = query.join(' ');
+      return throw 'ERROR IN THE QUERY: $sql WITH PARAMETES" $params';
+    }
   }
 
   String cleanRaw(String fmtString) {
@@ -50,9 +56,7 @@ class SQLquery {
 
     final sql = query.join(' ');
     print('==============================================================>');
-    print('1) QUERY: $sql');
-    print('2) PARAMETER: $params');
-    print('3) SQL EXECUTE QUERY: ${PostgreSQLFormat.substitute(sql, params)}');
+    print('SQL: ${PostgreSQLFormat.substitute(sql, params)}');
 
     final response = await connect.connection!.query(
       sql,
@@ -65,7 +69,11 @@ class SQLquery {
       final map = <String, dynamic>{};
       for (var i = 0; i < row.columnDescriptions.length; i++) {
         final column = row.columnDescriptions;
-        map[column[i].columnName] = row[i];
+        var key = column[i].columnName;
+        if (map.containsKey(key)) {
+          key = '$key' '_column$i';
+        }
+        map[key] = row[i];
       }
 
       if (transform != null) {
@@ -74,7 +82,6 @@ class SQLquery {
         result.add(map as T);
       }
     }
-
     return result;
   }
 
@@ -91,16 +98,25 @@ class SQLquery {
     }
 
     final sql = query.join(' ');
-    print('==============================================================>');
-    print('1) QUERY: $sql');
-    print('2) PARAMETER: $params');
-    print('3) SQL EXECUTE QUERY: ${PostgreSQLFormat.substitute(sql, params)}');
+    print('SQL: ${PostgreSQLFormat.substitute(sql, params)}');
 
     final result = await connect.connection!.execute(
       sql,
       substitutionValues: params,
     );
+    print('AFFECTED ROWS:: $result');
 
     return result;
+  }
+
+  void validateEnpty() {
+    if (query.isEmpty && selects.isEmpty) {
+      query.add('SELECT * FROM $table');
+    }
+
+    if (query.isEmpty && selects.isNotEmpty) {
+      final parameter = selects.join(', ');
+      query.add('SELECT $parameter FROM $table');
+    }
   }
 }