+++ test/mock/people_mock.dart
@@ -1,0 +1,8 @@
+class PeopleMock {
+  PeopleMock();
+
+  factory PeopleMock.fromJson(Map<String, dynamic> row) {
+    late final _ = row['vlues'];
+    return PeopleMock();
+  }
+}