+++ test/querys/insert_test.dart
@@ -1,0 +1,58 @@
+import 'package:database_query_builder/query_builder.dart';
+import 'package:database_query_builder/src/exec.dart';
+import 'package:test/test.dart';
+
+void main() {
+  test('Instance Table->insert', () {
+    final execute = DB.table('people').insert(
+      {
+        'first_name': 'new first name',
+        'last_name': 'new last name',
+        'age': 28,
+      },
+    );
+    expect(execute, isA<Execute>());
+  });
+
+  test('Table->insert->toSQL', () {
+    final execute = DB.table('people').insert(
+      {
+        'first_name': 'new first name',
+        'last_name': 'new last name',
+        'age': 28,
+      },
+    ).toSQL();
+
+    expect(
+      execute,
+      equals(
+        'INSERT INTO people (first_name, last_name, age) VALUES '
+        "('new first name', 'new last name', 28)",
+      ),
+    );
+  });
+
+  test('Table->insert->toSQL', () {
+    final execute = DB.table('people').insertAll([
+      {
+        'first_name': 'new first name 0',
+        'last_name': 'new last name 0',
+        'age': 28,
+      },
+      {
+        'first_name': 'new first name 1',
+        'last_name': 'new last name 1',
+        'age': 6,
+      },
+    ]).toSQL();
+
+    expect(
+      execute,
+      equals(
+        'INSERT INTO people (first_name, last_name, age) VALUES '
+        "('new first name 0', 'new last name 0', 28), "
+        "('new first name 1', 'new last name 1', 6);",
+      ),
+    );
+  });
+}